% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dr_plot_vfun}
\alias{dr_plot_vfun}
\title{Plot the fitted variance function}
\usage{
dr_plot_vfun(
  object,
  extrapolate = FALSE,
  col_curve = "#D6604D",
  col_pts = "#2166AC",
  xlab = "Design robustness",
  ylab = "Fitted tau^2",
  main = "Fitted variance function",
  ...
)
}
\arguments{
\item{object}{A fitted \code{drmeta} object.}

\item{extrapolate}{If TRUE, draw the curve over the whole interval
\eqn{[0, 1]} and shade the region beyond the observed support. Default
FALSE.}

\item{col_curve}{Colour of the fitted curve.}

\item{col_pts}{Colour of the study-level points.}

\item{xlab, ylab, main}{Axis labels and title.}

\item{...}{Further arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
Invisibly, the data frame of plotted grid values.
}
\description{
Draws the fitted residual between-study variance against the design index,
with study-level fitted variances overlaid. By default the curve is drawn
only over the observed range of the design index, since the fitted function
outside that range is an extrapolation.
}
\examples{
path <- system.file("extdata", "bcg_design_robustness.csv", package = "drmeta")
bcg <- utils::read.csv(path)
fit <- drmeta(yi = bcg[["yi"]], vi = bcg[["vi"]], dr = bcg[["dr"]])
dr_plot_vfun(fit)
}
