% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta.R
\name{coef.drmeta}
\alias{coef.drmeta}
\alias{vcov.drmeta}
\alias{fitted.drmeta}
\alias{residuals.drmeta}
\alias{print.drmeta}
\alias{summary.drmeta}
\alias{print.summary.drmeta}
\alias{drmeta-methods}
\alias{confint.drmeta}
\alias{logLik.drmeta}
\title{S3 methods for drmeta objects}
\usage{
\method{coef}{drmeta}(object, ...)

\method{vcov}{drmeta}(object, ...)

\method{fitted}{drmeta}(object, ...)

\method{residuals}{drmeta}(object, ...)

\method{print}{drmeta}(x, ...)

\method{summary}{drmeta}(object, ...)

\method{print}{summary.drmeta}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{confint}{drmeta}(object, parm = NULL, level = 0.95, ...)

\method{logLik}{drmeta}(object, REML = NULL, ...)
}
\arguments{
\item{object, x}{A fitted \code{drmeta} object.}

\item{...}{Ignored.}

\item{digits}{Number of significant digits.}

\item{parm}{Character or numeric selection of location parameters.}

\item{level}{Confidence level.}

\item{REML}{If TRUE, report the restricted log-likelihood.}
}
\value{
\code{coef} and \code{fitted} and \code{residuals} return numeric
vectors; \code{vcov} returns the location covariance matrix;
\code{confint} returns a matrix of confidence limits; \code{logLik}
returns an object of class \code{logLik} carrying \code{df} and
\code{nobs} attributes; \code{summary} returns an object of class
\code{summary.drmeta}; the print methods return their argument invisibly.
}
\description{
S3 methods for drmeta objects
}
