% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta_pkg.R
\docType{package}
\name{drmeta-package}
\alias{drmeta-package}
\title{drmeta: Design-Robust Meta-Analysis}
\description{
Fits the DR-Meta variance-function random-effects model (Hait, 2025),
where between-study heterogeneity is a monotone-decreasing function of
each study's design robustness index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/causalfragility-lab/drmeta}
  \item Report bugs at \url{https://github.com/causalfragility-lab/drmeta/issues}
}

}
\author{
\strong{Maintainer}: Subir Hait \email{haitsubi@msu.edu} (\href{https://orcid.org/0009-0004-9871-9677}{ORCID})

}
