% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{normalize_01}
\alias{normalize_01}
\title{Rescale a numeric vector to the unit interval}
\usage{
normalize_01(x)
}
\arguments{
\item{x}{A numeric vector. Missing values are ignored during rescaling and
preserved in the result.}
}
\value{
A numeric vector of the same length rescaled to \eqn{[0, 1]}. If all
non-missing values are equal, a zero vector is returned.
}
\description{
Rescale a numeric vector to the unit interval
}
\examples{
normalize_01(c(2, 5, 8))
normalize_01(c(1, 1, 1))
}
