% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta.R
\name{residuals.drmeta}
\alias{residuals.drmeta}
\title{Residuals for a drmeta Object}
\usage{
\method{residuals}{drmeta}(object, type = c("raw", "standardised"), ...)
}
\arguments{
\item{object}{A fitted \code{"drmeta"} object.}

\item{type}{\code{"raw"} (default) or \code{"standardised"}.}

\item{...}{Ignored.}
}
\description{
Residuals for a drmeta Object
}
