% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta.R
\name{logLik.drmeta}
\alias{logLik.drmeta}
\title{Log-Likelihood for a drmeta Object}
\usage{
\method{logLik}{drmeta}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{A fitted \code{"drmeta"} object.}

\item{REML}{Logical.  If \code{TRUE}, returns the REML log-likelihood.
Default \code{FALSE} (ML).}

\item{...}{Ignored.}
}
\description{
Log-Likelihood for a drmeta Object
}
