% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta.R
\name{print.drmeta}
\alias{print.drmeta}
\title{Print Method for drmeta Objects}
\usage{
\method{print}{drmeta}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted \code{"drmeta"} object.}

\item{digits}{Number of significant digits.  Default 4.}

\item{...}{Ignored.}
}
\value{
Invisibly returns the original \code{drmeta} object \code{x},
unchanged. This function is called for its side effect of printing a
formatted summary of the fitted DR-Meta model to the console.
}
\description{
Print Method for drmeta Objects
}
