% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.drmeta_bootstrap_gamma}
\alias{print.drmeta_bootstrap_gamma}
\title{Print a parametric-bootstrap scale-gradient test}
\usage{
\method{print}{drmeta_bootstrap_gamma}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class drmeta_bootstrap_gamma.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}. Called for its printed output.
}
\description{
Print a parametric-bootstrap scale-gradient test
}
