% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmeta.R
\name{summary.drmeta}
\alias{summary.drmeta}
\title{Summary Method for drmeta Objects}
\usage{
\method{summary}{drmeta}(object, digits = 4, ...)
}
\arguments{
\item{object}{A fitted \code{"drmeta"} object.}

\item{digits}{Number of significant digits.  Default 4.}

\item{...}{Ignored.}
}
\description{
Summary Method for drmeta Objects
}
